pyproject.toml
[project]
name = "teacher-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "bedrock-agentcore>=1.1.2",
    "fastapi>=0.127.0",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "strands-agents>=1.20.0",
    "strands-agents-tools>=0.2.18",
    "uvicorn[standard]>=0.40.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.10",
    "strands-agents[ollama]>=1.20.0",
]
